+++ frontend/src/lib/components/graph/renderers/nodes/ForLoopEndNode.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
 	import VirtualItem from '$lib/components/flows/map/VirtualItem.svelte'
+	// @ts-ignore
 	import { Handle, NodeToolbar, Position, type NodeProps } from '@xyflow/svelte'
 
 	type $$Props = NodeProps