+++ frontend/src/lib/components/graph/renderers/nodes/ForLoopStartNode.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 	import DarkModeObserver from '$lib/components/DarkModeObserver.svelte'
 	import VirtualItem from '$lib/components/flows/map/VirtualItem.svelte'
+	// @ts-ignore
 	import { Handle, NodeToolbar, Position, type NodeProps } from '@xyflow/svelte'
 
 	type $$Props = NodeProps